move run/android/resolveLaunchProps.ts -> start/platforms/android/resolveLaunchProps.ts
